WIP: Allow make<> detection to work in retail with no overhead
The idea is to reuse an existing virtual method in base_implements.
The implementations all leave the method unimplemented, and
heap_implements finally implements it by forwarding to a nonvirtual
equivalent.

This commit contains the basic idea, but it still doesn't build.
Need to chase down dependencies. (And validate that it even works.)